+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/Constants.java
@@ -111,6 +111,8 @@ public final class Constants {
 	public static final String SITUATION_NAME = "situation";
 	public static final String CATEGORY_NAME = "category";
 	public static final String OCCUPANT_NAME = "occupant";
+
+	public static final String OUTCOME_SPOTTING = "outcome_spotting";
 	public static final String REPORTING_DATA_SURVEY_VALIDATION_NAME = REPORTING_DATA_PREFIX_NAME + SURVEY_VALIDATION_DATE_NAME;
 	
 	// ------ Pairwise variables