+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/dataprocessing/TCMModuleEnum.java
@@ -3,7 +3,8 @@ package fr.insee.kraftwerk.core.dataprocessing;
 public enum TCMModuleEnum {
     TCM_THL_DET
     ,TCM_THL_SIMPLE
-    ,TCM_THL_LDV
+    ,TCM_THL_LDV,
+    TCM_THL_CDM
     ,TCM_ACT_PRINC
     ,TCM_ACT_ANTE
     ,TCM_ACT_BIS